projects/shared/upwork-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "upwork-mcp"
version = "1.0.0"
description = "MCP server for Upwork job discovery, client analysis, and proposal assistance"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "William Marceau", email = "wmarceau@marceausolutions.com"}
]
keywords = ["mcp", "claude", "upwork", "freelance", "jobs", "proposals", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "python-upwork-oauth2>=3.0.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/MarceauSolutions/upwork-mcp"
Repository = "https://github.com/MarceauSolutions/upwork-mcp"

[project.scripts]
upwork-mcp = "upwork_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/upwork_mcp"]
